Refactored buildings list and individual building view into buildings folder
