+++ energy-app-react/src/ComparisonPage.js
@@ -3,8 +3,9 @@ import { FlatList, AppRegistry, SectionList, StyleSheet, View, Text, Image, WebV
 import { StackNavigator, SafeAreaView } from 'react-navigation';
 import { List, Card, ListItem, Button, Avatar } from 'react-native-elements';
 import { Table, TableWrapper, Row, Rows, Col, Cols, Cell } from 'react-native-table-component';
-import buildings from './Buildings';
-import IndividualBuiding from './IndividualBuilding';
+
+import buildings from './buildings/Buildings';
+import IndividualBuiding from './buildings/IndividualBuilding';
 import { GetStyle } from './styling/Themes';
 import CurrTheme from './styling/CurrentTheme';
 import BuildingComparison from './BuildingComparison';
