+++ energy-app-react/src/buildings/Buildings.js
