+++ energy-app-react/src/energymap/EnergyMapView.js
@@ -6,9 +6,9 @@ import MapView, { PROVIDER_GOOGLE, Polygon, Callout, Marker } from 'react-native
 
 import { connect } from 'react-redux';
 import MapCallout from './MapCallout';
-import IndividualBuilding from './../IndividualBuilding';
-import BuildingStack from './../BuildingListView';
-import buildings from './../Buildings'
+import IndividualBuilding from './../buildings/IndividualBuilding';
+import BuildingStack from './../buildings/BuildingListView';
+import buildings from './../buildings/Buildings'
 import { getCurrentBuildingUtilityConsumption, getUtilitiesList } from './../helpers/ApiWrappers.js';
 import TopUtilities from './EnergyMapUtilityButtons';
 import EnergyMapTimestamp from './EnergyMapTimestamp';
